fix: handle race in inline azVolume deletion
